# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Library Portfolio API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: books
    description: Book management
  - name: stats
    description: Library statistics
  - name: journal
    description: Personal journal entries
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /books/bengali:
    get:
      operationId: listBengaliBooks
      tags: [books]
      summary: List all Bengali language books
      responses:
        "200":
          description: Bengali books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"

  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List all books
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search by title or author
        - in: query
          name: genre
          schema:
            type: string
          description: Filter by genre
        - in: query
          name: status
          schema:
            type: string
            enum: [read, reading, want_to_read]
          description: Filter by reading status
        - in: query
          name: language
          schema:
            type: string
            enum: [english, bengali]
          description: Filter by language
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: createBook
      tags: [books]
      summary: Add a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookBody"
      responses:
        "201":
          description: Created book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get a book by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBook
      tags: [books]
      summary: Update a book
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookBody"
      responses:
        "200":
          description: Updated book
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /journal:
    get:
      operationId: listJournalEntries
      tags: [journal]
      summary: List all journal entries for the logged-in user
      responses:
        "200":
          description: Journal entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalEntry"
    post:
      operationId: createJournalEntry
      tags: [journal]
      summary: Create a new journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJournalEntryBody"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"

  /journal/{id}:
    get:
      operationId: getJournalEntry
      tags: [journal]
      summary: Get a single journal entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateJournalEntry
      tags: [journal]
      summary: Update a journal entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJournalEntryBody"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteJournalEntry
      tags: [journal]
      summary: Delete a journal entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get library statistics
      responses:
        "200":
          description: Library stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryStats"

  /books/genres:
    get:
      operationId: listGenres
      tags: [books]
      summary: List all unique genres
      responses:
        "200":
          description: List of genres
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /books/recent:
    get:
      operationId: listRecentBooks
      tags: [books]
      summary: List recently added books
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Max results (default 5)
      responses:
        "200":
          description: Recent books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"

  /books/favorites:
    get:
      operationId: listFavoriteBooks
      tags: [books]
      summary: List top-rated (favorite) books
      responses:
        "200":
          description: Favorite books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        genres:
          type: array
          items:
            type: string
        summary:
          type: ["string", "null"]
        quotes:
          type: array
          items:
            type: string
        review:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
          description: Rating from 1 to 5
        status:
          type: string
          enum: [read, reading, want_to_read]
        language:
          type: string
          enum: [english, bengali]
          description: Language of the book
        coverUrl:
          type: ["string", "null"]
        publishedYear:
          type: ["integer", "null"]
        pageCount:
          type: ["integer", "null"]
        currentPage:
          type: ["integer", "null"]
          description: Current page the reader is on
        readingDeadline:
          type: ["string", "null"]
          description: Target date to finish the book (YYYY-MM-DD)
        format:
          type: ["string", "null"]
          enum: [pdf, physical, null]
          description: How the book was read (pdf or physical copy)
        driveLink:
          type: ["string", "null"]
          description: Link to cloud storage where the book file is stored (Google Drive, Dropbox, etc)
        completedAt:
          type: ["string", "null"]
          description: Date the book was actually completed (YYYY-MM-DD)
        isOwned:
          type: boolean
          description: Whether the physical book is at home
        wantToBuy:
          type: boolean
          description: Whether the user wants to purchase this book
        isFavorite:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - author
        - genres
        - quotes
        - status
        - language
        - isOwned
        - wantToBuy
        - isFavorite
        - createdAt
        - updatedAt

    CreateBookBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        genres:
          type: array
          items:
            type: string
        summary:
          type: ["string", "null"]
        quotes:
          type: array
          items:
            type: string
        review:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        status:
          type: string
          enum: [read, reading, want_to_read]
          default: want_to_read
        language:
          type: string
          enum: [english, bengali]
          default: english
        coverUrl:
          type: ["string", "null"]
        publishedYear:
          type: ["integer", "null"]
        pageCount:
          type: ["integer", "null"]
        currentPage:
          type: ["integer", "null"]
        readingDeadline:
          type: ["string", "null"]
        format:
          type: ["string", "null"]
          enum: [pdf, physical, null]
        driveLink:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        isOwned:
          type: boolean
          default: false
        wantToBuy:
          type: boolean
          default: false
        isFavorite:
          type: boolean
          default: false
      required:
        - title
        - author

    UpdateBookBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        genres:
          type: array
          items:
            type: string
        summary:
          type: ["string", "null"]
        quotes:
          type: array
          items:
            type: string
        review:
          type: ["string", "null"]
        rating:
          type: ["integer", "null"]
        status:
          type: string
          enum: [read, reading, want_to_read]
        language:
          type: string
          enum: [english, bengali]
        coverUrl:
          type: ["string", "null"]
        publishedYear:
          type: ["integer", "null"]
        pageCount:
          type: ["integer", "null"]
        currentPage:
          type: ["integer", "null"]
        readingDeadline:
          type: ["string", "null"]
        format:
          type: ["string", "null"]
          enum: [pdf, physical, null]
        driveLink:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        isOwned:
          type: boolean
        wantToBuy:
          type: boolean
        isFavorite:
          type: boolean

    LibraryStats:
      type: object
      properties:
        totalBooks:
          type: integer
        booksRead:
          type: integer
        booksReading:
          type: integer
        booksWantToRead:
          type: integer
        averageRating:
          type: ["number", "null"]
        totalPages:
          type: integer
        genreBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/GenreCount"
        topAuthors:
          type: array
          items:
            $ref: "#/components/schemas/AuthorCount"
        booksPerMonth:
          type: array
          items:
            $ref: "#/components/schemas/MonthCount"
        ratingDistribution:
          type: array
          items:
            $ref: "#/components/schemas/RatingCount"
        languageBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/LangCount"
        formatBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/FormatCount"
        pagesReadTotal:
          type: integer
        avgPagesPerBook:
          type: integer
        booksPerYear:
          type: array
          items:
            $ref: "#/components/schemas/YearCount"
      required:
        - totalBooks
        - booksRead
        - booksReading
        - booksWantToRead
        - totalPages
        - genreBreakdown
        - topAuthors
        - booksPerMonth
        - ratingDistribution
        - languageBreakdown
        - formatBreakdown
        - pagesReadTotal
        - avgPagesPerBook
        - booksPerYear

    GenreCount:
      type: object
      properties:
        genre:
          type: string
        count:
          type: integer
      required:
        - genre
        - count

    AuthorCount:
      type: object
      properties:
        author:
          type: string
        count:
          type: integer
      required:
        - author
        - count

    MonthCount:
      type: object
      properties:
        month:
          type: string
        count:
          type: integer
      required:
        - month
        - count

    RatingCount:
      type: object
      properties:
        rating:
          type: integer
        count:
          type: integer
      required:
        - rating
        - count

    LangCount:
      type: object
      properties:
        language:
          type: string
        count:
          type: integer
      required:
        - language
        - count

    FormatCount:
      type: object
      properties:
        format:
          type: string
        count:
          type: integer
      required:
        - format
        - count

    YearCount:
      type: object
      properties:
        year:
          type: integer
        count:
          type: integer
      required:
        - year
        - count

    JournalEntry:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        mood:
          type: string
          enum: [happy, reflective, inspired, melancholic, neutral]
        domain:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        bookId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - mood
        - tags
        - createdAt
        - updatedAt

    CreateJournalEntryBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        mood:
          type: string
          enum: [happy, reflective, inspired, melancholic, neutral]
          default: neutral
        domain:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        bookId:
          type: ["integer", "null"]
      required:
        - title

    UpdateJournalEntryBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        mood:
          type: string
          enum: [happy, reflective, inspired, melancholic, neutral]
        domain:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        bookId:
          type: ["integer", "null"]