# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Library Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { BookFormat } from "./bookFormat";
import type { BookLanguage } from "./bookLanguage";
import type { BookStatus } from "./bookStatus";

export interface Book {
  id: number;
  title: string;
  author: string;
  genres: string[];
  /** @nullable */
  summary?: string | null;
  quotes: string[];
  /** @nullable */
  review?: string | null;
  /**
   * Rating from 1 to 5
   * @nullable
   */
  rating?: number | null;
  status: BookStatus;
  /** Language of the book */
  language: BookLanguage;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  publishedYear?: number | null;
  /** @nullable */
  pageCount?: number | null;
  /**
   * Current page the reader is on
   * @nullable
   */
  currentPage?: number | null;
  /**
   * Target date to finish the book (YYYY-MM-DD)
   * @nullable
   */
  readingDeadline?: string | null;
  /**
   * How the book was read (pdf or physical copy)
   * @nullable
   */
  format?: BookFormat;
  /**
   * Link to cloud storage where the book file is stored (Google Drive, Dropbox, etc)
   * @nullable
   */
  driveLink?: string | null;
  /**
   * Date the book was actually completed (YYYY-MM-DD)
   * @nullable
   */
  completedAt?: string | null;
  /** Whether the physical book is at home */
  isOwned: boolean;
  /** Whether the user wants to purchase this book */
  wantToBuy: boolean;
  isFavorite: boolean;
  createdAt: Date;
  updatedAt: Date;
}