# lib/api-zod/src/generated/types/createBookBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Library Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookBodyFormat } from "./createBookBodyFormat";
import type { CreateBookBodyLanguage } from "./createBookBodyLanguage";
import type { CreateBookBodyStatus } from "./createBookBodyStatus";

export interface CreateBookBody {
  title: string;
  author: string;
  genres?: string[];
  /** @nullable */
  summary?: string | null;
  quotes?: string[];
  /** @nullable */
  review?: string | null;
  /** @nullable */
  rating?: number | null;
  status?: CreateBookBodyStatus;
  language?: CreateBookBodyLanguage;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  publishedYear?: number | null;
  /** @nullable */
  pageCount?: number | null;
  /** @nullable */
  currentPage?: number | null;
  /** @nullable */
  readingDeadline?: string | null;
  /** @nullable */
  format?: CreateBookBodyFormat;
  /** @nullable */
  driveLink?: string | null;
  /** @nullable */
  completedAt?: string | null;
  isOwned?: boolean;
  wantToBuy?: boolean;
  isFavorite?: boolean;
}