# lib/api-zod/src/generated/types/createJournalEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Library Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateJournalEntryBodyMood } from "./createJournalEntryBodyMood";

export interface CreateJournalEntryBody {
  title: string;
  content?: string;
  mood?: CreateJournalEntryBodyMood;
  /** @nullable */
  domain?: string | null;
  tags?: string[];
  /** @nullable */
  bookId?: number | null;
}