# lib/api-zod/src/generated/types/journalEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Library Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { JournalEntryMood } from "./journalEntryMood";

export interface JournalEntry {
  id: number;
  title: string;
  content: string;
  mood: JournalEntryMood;
  /** @nullable */
  domain?: string | null;
  tags: string[];
  /** @nullable */
  bookId?: number | null;
  createdAt: Date;
  updatedAt: Date;
}