# lib/api-zod/src/generated/types/updateJournalEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Library Portfolio API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateJournalEntryBodyMood } from "./updateJournalEntryBodyMood";

export interface UpdateJournalEntryBody {
  title?: string;
  content?: string;
  mood?: UpdateJournalEntryBodyMood;
  /** @nullable */
  domain?: string | null;
  tags?: string[];
  /** @nullable */
  bookId?: number | null;
}